feat: add initial support for tags sync
refactor navigation w.r.t. collection screen context [wip],
add support for multiple app icons, and a couple of other things.
really need to stop squashing stuff and messing up the history

+++ composeApp/src/commonMain/kotlin/com/sakethh/linkora/ui/screens/settings/section/data/DataSettingsScreen.kt
@@ -736,14 +736,18 @@ fun DataSettingsScreen() {
         operationTitle = Localization.Key.SyncingDataLabel.rememberLocalizedString(),
         operationDesc = Localization.Key.InitiateManualSyncDescAlt.rememberLocalizedString()
     )
-    DeleteDialogBox(
-        deleteDialogBoxParam = DeleteDialogBoxParam(
-            shouldDialogBoxAppear = shouldDeleteEntireDialogBoxAppear,
-            deleteDialogBoxType = DeleteDialogBoxType.REMOVE_ENTIRE_DATA,
-            onDeleteClick = { onCompletion, deleteEverythingFromRemote ->
-                dataSettingsScreenVM.deleteEntireDatabase(deleteEverythingFromRemote, onCompletion)
-            })
-    )
+    if (shouldDeleteEntireDialogBoxAppear.value){
+        DeleteDialogBox(
+            deleteDialogBoxParam = DeleteDialogBoxParam(
+                onDismiss = {
+                    shouldDeleteEntireDialogBoxAppear.value = false
+                },
+                deleteDialogBoxType = DeleteDialogBoxType.REMOVE_ENTIRE_DATA,
+                onDeleteClick = { onCompletion, deleteEverythingFromRemote ->
+                    dataSettingsScreenVM.deleteEntireDatabase(deleteEverythingFromRemote, onCompletion)
+                })
+        )
+    }
     if (showDuplicateDeleteDialogBox.value) {
         AlertDialog(onDismissRequest = {}, content = {
             Column(
